Extract framework-neutral OpenAPI core
