+++ packages/client/src/utils/openapi-normalizer.ts
@@ -1,15 +1,10 @@
+import type { OpenAPISpec, Operation, Parameter, PathItem, Reference, RequestBody, Response, SecurityRequirement, Server } from '../types/openapi';
 import {
-  OpenAPISpec,
-  Operation,
-  Parameter,
-  PathItem,
-  Reference,
-  RequestBody,
-  Response,
-  SecurityRequirement,
-  Server,
-} from '../types/openapi';
-import { OpenAPIParser } from './openapi-parser';
+  normalizeOperation as coreNormalizeOperation,
+  resolveObject as coreResolveObject,
+  resolvePathItem as coreResolvePathItem,
+  resolveServerVariables as coreResolveServerVariables,
+} from '../../../core/src/openapi-normalizer';
 
 export interface NormalizedOperation {
   path: string;
@@ -24,58 +19,10 @@ export interface NormalizedOperation {
 }
 
 export function resolveObject<T>(spec: OpenAPISpec, value: T | Reference | undefined): T | undefined {
-  if (!value) return undefined;
-  return OpenAPIParser.isReference(value)
-    ? (OpenAPIParser.resolveReference(spec, value.$ref) as T)
-    : value as T;
+  return coreResolveObject(spec, value) as T | undefined;
 }
-
-export function resolvePathItem(spec: OpenAPISpec, path: string): PathItem {
-  const raw = spec.paths[path] as PathItem | Reference | undefined;
-  if (!raw) throw new Error(`Path not found: ${path}`);
-  return resolveObject<PathItem>(spec, raw)!;
-}
-
-export function resolveServerVariables(server: Server, values: Record<string, string> = {}): string {
-  return server.url.replace(/\{([^}]+)\}/g, (_, name: string) => {
-    const variable = server.variables?.[name];
-    const value = values[name] ?? variable?.default;
-    if (value === undefined) throw new Error(`Missing server variable: ${name}`);
-    if (variable?.enum?.length && !variable.enum.includes(value)) {
-      throw new Error(`Invalid value for server variable ${name}: ${value}`);
-    }
-    return value;
-  });
-}
-
+export function resolvePathItem(spec: OpenAPISpec, path: string): PathItem { return coreResolvePathItem(spec, path) as PathItem; }
+export function resolveServerVariables(server: Server, values: Record<string, string> = {}): string { return coreResolveServerVariables(server, values); }
 export function normalizeOperation(spec: OpenAPISpec, path: string, method: string): NormalizedOperation {
-  const pathItem = resolvePathItem(spec, path);
-  const operation = pathItem[method.toLowerCase() as keyof PathItem] as Operation | undefined;
-  if (!operation || typeof operation !== 'object' || !('responses' in operation)) {
-    throw new Error(`Operation not found: ${method.toUpperCase()} ${path}`);
-  }
-
-  const merged = [...(pathItem.parameters || []), ...(operation.parameters || [])]
-    .map((parameter) => resolveObject<Parameter>(spec, parameter))
-    .filter((parameter): parameter is Parameter => !!parameter);
-  const parameters = new Map<string, Parameter>();
-  for (const parameter of merged) parameters.set(`${parameter.in}:${parameter.name}`, parameter);
-
-  const responses: Record<string, Response> = {};
-  for (const [status, response] of Object.entries(operation.responses || {})) {
-    const resolved = resolveObject<Response>(spec, response);
-    if (resolved) responses[status] = resolved;
-  }
-
-  return {
-    path,
-    method: method.toUpperCase(),
-    pathItem,
-    operation,
-    parameters: [...parameters.values()],
-    requestBody: resolveObject<RequestBody>(spec, operation.requestBody),
-    responses,
-    servers: operation.servers || pathItem.servers || spec.servers || [],
-    security: operation.security !== undefined ? operation.security : (spec.security || []),
-  };
+  return coreNormalizeOperation(spec, path, method) as NormalizedOperation;
 }
